% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrediction.R
\name{createModelPrediction}
\alias{createModelPrediction}
\title{Create Prediction}
\usage{
createModelPrediction(object, data, target = NULL, predict.fun = NULL)
}
\arguments{
\item{object}{[any] \cr
Either a \code{\link[mlr]{WrappedModel}} or a \code{\link[mlr]{ResampleResult}}.
You can also provide any other trained model (however, this is experimental and you need to provide an appropriate \code{predict.fun}).}

\item{data}{[\code{data.frame}] \cr
The data whose features will be permuted in order to measure their importance.
If \code{object} is of class \code{ResampleResult}, you should use the corresponding data on which the whole resampling was performed. In any other case, you might use some independent test data that was not used to fit the model (although you could also use the train data here).}

\item{target}{[\code{character(1)}] \cr
Only needed if \code{object} is not of class \code{WrappedModel} or \code{ResampleResult}.
Name of the target feature to be predicted.}

\item{predict.fun}{[\code{function}] \cr
Only needed if \code{object} is not of class \code{WrappedModel} or \code{ResampleResult}.
The signature must be \code{function(object, newdata)} and the function should always return a vector of predictions.
The default \code{NULL} internally uses \code{predict(object, newdata = newdata)}.}
}
\description{
Make a predicition object.
}
