% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureImportance.R
\name{featureImportance}
\alias{featureImportance}
\title{Feature Importance}
\usage{
featureImportance(object, data, target = NULL, measures, minimize = NULL,
  features, n.feat.perm = 10, local = FALSE, predict.fun = NULL,
  contrast.fun = NULL, ...)
}
\arguments{
\item{object}{[\code{\link[mlr]{WrappedModel}} | \code{\link[mlr]{ResampleResult}}] \cr
Currently either the trained \code{\link[mlr]{WrappedModel}} or a \code{\link[mlr]{ResampleResult}}.}

\item{data}{[\code{data.frame}] \cr
The data whose features will be shuffled in order to measure the performance drop.
If \code{object} is of class \cr
\code{ResampleResult}, you should use the corresponding data on which the resampling was performed \cr
\code{WrappedModel}, you should use some independent test data that was not used to fit the model (although you could also use the train data here).}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}}] \cr
Performance measure(s) used to measure the drop in performance.}

\item{features}{[\code{character}] \cr
A vector of features for which the importance should be computed.}

\item{n.feat.perm}{[\code{numeric(1)}] \cr
The number of Monte-Carlo iterations, e.g. number of permutations of the feature(s) to compute the feature importance.
The default is 10.}

\item{local}{[\code{logical(1)}] \cr
Should observation-wise importance be computed?
Note that not all measures support this (e.g. one can not compute the AUC for one observation).
The default is FALSE.}

\item{...}{Not used.}
}
\description{
Measures the feature importance by drop in performance.
}
