% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureImportanceLearner.R
\name{featureImportanceLearner}
\alias{featureImportanceLearner}
\title{Feature Importance}
\usage{
featureImportanceLearner(
  learner,
  task,
  resampling,
  measures = mlr::getDefaultMeasure(task),
  features = as.list(mlr::getTaskFeatureNames(task)),
  n.feat.perm = 50,
  ...
)
}
\arguments{
\item{learner}{[\code{\link[mlr]{Learner}}] \cr
The learner.}

\item{task}{(\link[mlr]{Task})\cr
The task.}

\item{resampling}{(\link[mlr]{ResampleDesc} or \link[mlr]{ResampleInstance})\cr
Resampling strategy.
If a description is passed, it is instantiated automatically.}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}} | \code{function} | list of \code{function}] \cr
Performance measure(s) used to measure the model performance.
Can also be a named list of function with signature \code{function(y, pred)},
where \code{y} and \code{pred} are vectors containing the true and the predicted values of the target.}

\item{features}{[list of \code{character}] \cr
A list where each element contains the names of at least one feature for which the permutation importance should be computed.
If a list element contains two or more features, they will be permuted block-wise (without breaking the relationship between those features).}

\item{n.feat.perm}{[\code{numeric(1)}] \cr
The number of permutations of the feature(s) used to compute the feature importance.
The default is 50.
If set to NULL, the cartesian product is used.}

\item{...}{Not used.}
}
\description{
Measures the feature importance by drop in performance.
}
