% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureFeatureImportance.R
\name{measureFeatureImportance}
\alias{measureFeatureImportance}
\title{Measures the Feature Importance
Measures the drop in performance between permuted and unpermuted features w.r.t. a function.}
\usage{
measureFeatureImportance(permuted.perf, unpermuted.perf,
  importance.fun = NULL)
}
\arguments{
\item{permuted.perf}{a vector of the performance(s) when a feature was shuffled}

\item{unpermuted.perf}{a vector of the true performance(s)}

\item{measures}{the performance measures that have been used: if big values for the measure are better, the drop in performance is true - permuted (negative "drop" values are performance "gains")}
}
\description{
Measures the Feature Importance
Measures the drop in performance between permuted and unpermuted features w.r.t. a function.
}
