% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{models}
\alias{models}
\title{models selector for wrapper importance methods}
\usage{
models(name, metric, mod_args = list())
}
\description{
models selector for wrapper importance methods
}
\keyword{internal}
