% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{op_util}
\alias{op_util}
\title{utility function for formatting single operation}
\usage{
op_util(x, operator, n = 1)
}
\description{
utility function for formatting single operation
}
\keyword{internal}
