% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimp.R
\name{pimp}
\alias{pimp}
\title{Permutation Importance}
\usage{
pimp(learner, task, resampling, measures, features = NULL,
  n.feat.perm = 1, n.target.perm = 1)
}
\arguments{
\item{learner}{[\code{\link[mlr]{Learner}}] \cr
The learner.}

\item{task}{(\link{Task})\cr
The task.}

\item{resampling}{(\link{ResampleDesc} or \link{ResampleInstance})\cr
Resampling strategy.
If a description is passed, it is instantiated automatically.}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}} | \code{function} | list of \code{function}] \cr
Performance measure(s) used to measure the model performance.
Can also be a named list of function with signature \code{function(y, pred)},
where \code{y} and \code{pred} are vectors containing the true and the predicted values of the target.}

\item{features}{[list of \code{character}] \cr
A list where each element contains the names of at least one feature for which the permutation importance should be computed.
If a list element contains two or more features, they will be permuted block-wise (without breaking the relationship between those features).}

\item{n.feat.perm}{[\code{numeric(1)}] \cr
The number of permutations of the feature(s) used to compute the feature importance.
The default is 50.
If set to NULL, the cartesian product is used.}

\item{n.target.perm}{[\code{numeric(1)}] \cr
The number of permutations of the target used to compute a null distribution for the importance values.}
}
\description{
Implementation of PIMP (see Altmann et al. (2010))
}
\references{
Altmann, A., Tolosi, L., Sander, O., & Lengauer, T. (2010).
Permutation importance: a corrected feature importance measure.
Bioinformatics, 26(10), 1340-1347.
}
