% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_last_n}
\alias{remove_last_n}
\title{Remove Last N Characters}
\usage{
remove_last_n(str, n = 0, missing = "")
}
\arguments{
\item{str}{the string or vector of strings to be truncated}

\item{n}{the number of characters at the end to remove}

\item{missing}{optional sequence to use if an entry is truncated away (default = '')}
}
\value{
the string or strings, reduced by n characters
}
\description{
utility function to remove n characters from the end of a character string or strings.
Extends functional tools available in stringr package for this common operation.  Also allows for
a set of characters (missing='' by default) to be specified to show that the string was over-truncated.
}
\examples{
remove_last_n('a sequence #', 2)
remove_last_n(c('a couple+','of sequences+'), 1, '...')
}
