% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{GOFFI}
\alias{GOFFI}
\title{Goodness of Fit Feature Importance}
\usage{
GOFFI(
  X,
  y,
  metric = c("residual", "AIC", "BIC"),
  mod_type = c("lm", "glm"),
  ...
)
}
\arguments{
\item{X}{the features of the dataset}

\item{y}{the target}

\item{metric}{used to determine importance ranking}

\item{mod_type}{type of model used for y ~ X ( lm, glm )}

\item{...}{other optional arguments to the model}
}
\value{
a named vector with the relative importance for each feature in X
}
\description{
This wrapper method for feature importance ranking calculates an importance
metric for each feature corresponding to the sum of ratios of model coefficients (or weights)
to the model goodness of fit for a set of fitted models which each exclude one feature.
}
\examples{
X <- mtcars[,-1]
y <- mtcars[,1]
GOFFI(X, y)
}
