% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureFeatureImportance.R
\name{measureFeatureImportance}
\alias{measureFeatureImportance}
\title{Measures the Feature Importance
Measures the drop in performance between permuted and unpermuted features w.r.t. a function.}
\usage{
measureFeatureImportance(permuted.perf, unpermuted.perf,
  importance.fun = NULL)
}
\arguments{
\item{permuted.perf}{a vector of the performance(s) when a feature was shuffled}

\item{unpermuted.perf}{a vector of the true performance(s)}

\item{importance.fun}{[\code{function}] \cr
Function with signature \code{function(permuted, unpermuted)} which defines how \code{permuted} and \code{unpermuted} are aggregated to a feature importance measure.
The function takes the result of \code{\link{measurePerformance}} as input for \code{permuted} and \code{unpermuted}.
The default \code{NULL} internally uses \code{unpermuted - permuted} which refers to the drop in performance if the measure is to be maximized (i.e., if higher values of the measure refer to better performance).}
}
\description{
Measures the Feature Importance
Measures the drop in performance between permuted and unpermuted features w.r.t. a function.
}
