% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurePerformance.R
\name{measurePerformance}
\alias{measurePerformance}
\title{Measure Performance}
\usage{
measurePerformance(
  object,
  data,
  target = NULL,
  measures,
  local = FALSE,
  predict.fun = NULL
)
}
\arguments{
\item{object}{[any] \cr
Either a \code{\link[mlr]{WrappedModel}} or a \code{\link[mlr]{ResampleResult}}.
You can also provide any other trained model (however, this is experimental and you need to provide an appropriate \code{predict.fun}).}

\item{data}{[\code{data.frame}] \cr
The data whose features will be permuted in order to measure their importance.
If \code{object} is of class \code{ResampleResult}, you should use the corresponding data on which the whole resampling was performed. In any other case, you might use some independent test data that was not used to fit the model (although you could also use the train data here).}

\item{target}{[\code{character(1)}] \cr
Only needed if \code{object} is not of class \code{WrappedModel} or \code{ResampleResult}.
Name of the target feature to be predicted.}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}} | \code{function} | list of \code{function}] \cr
Performance measure(s) used to measure the model performance.
Can also be a named list of function with signature \code{function(y, pred)},
where \code{y} and \code{pred} are vectors containing the true and the predicted values of the target.}

\item{local}{[\code{logical(1)}] \cr
Should the performance (or the feature importance) be computed observation-wise?
Note that not all measures support this (e.g. one can not compute the AUC for one observation).
The default is FALSE.}

\item{predict.fun}{[\code{function}] \cr
Only needed if \code{object} is not of class \code{WrappedModel} or \code{ResampleResult}.
The signature must be \code{function(object, newdata)} and the function should always return a vector of predictions.
The default \code{NULL} internally uses \code{predict(object, newdata = newdata)}.}
}
\description{
Measures the performance on passed data.
}
