% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImportance.R
\name{plotImportance}
\alias{plotImportance}
\title{Plot function for feature importance}
\usage{
plotImportance(
  importance,
  feature,
  mid,
  individual = FALSE,
  hline = TRUE,
  grid.points = TRUE
)
}
\arguments{
\item{importance}{[\code{featureImportance | data.frame}] \cr
Object of class \code{featureImportance} or a \code{data.frame} containing the same information as the result of the \code{featureImportance}'s \code{$importance} slot.}

\item{feature}{[\code{character(1)}] \cr
The feature for which the PI or ICI curves should be visualized.}

\item{mid}{[\code{character(1)}] \cr
The measure name used for computing the feature importance.}

\item{individual}{[\code{logical(1)}] \cr
If \code{TRUE}, the ICI curves are additionally plotted. Otherwise only the PI curve is plotted.}

\item{hline}{[\code{logical(1)}] \cr
Whether a horizontal line should be plotted whose y-position refers to the feature importance of the considered feature.}

\item{grid.points}{[\code{logical(1)}] \cr
Whether points should be plotted whose x-position refer to the feature values used to compute the curves.}
}
\description{
Allows to visualize the PI or ICI curves.
}
\references{
Casalicchio, G., Molnar, M., & Bischl, B. (2018).
Visualizing the Feature Importance for Black Box Models.
arXiv preprint arXiv:1804.06620 (2018).
}
