% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapleyImportance.R
\name{shapleyImportance}
\alias{shapleyImportance}
\title{Shapley Importance}
\usage{
shapleyImportance(object, data, features, target, bound.size = NULL,
  n.feat.perm = 50, n.shapley.perm = 120, measures, predict.fun = NULL,
  value.function = calculateValueFunctionImportance)
}
\arguments{
\item{object}{[any] \cr
Either a \code{\link[mlr]{WrappedModel}} or a \code{\link[mlr]{ResampleResult}}.
You can also provide any other trained model (however, this is experimental and you need to provide an appropriate \code{predict.fun}).}

\item{data}{[\code{data.frame}] \cr
The data whose features will be permuted in order to measure their importance.
If \code{object} is of class \code{ResampleResult}, you should use the corresponding data on which the whole resampling was performed. In any other case, you might use some independent test data that was not used to fit the model (although you could also use the train data here).}

\item{features}{[\code{character}] \cr
The feature(s) for which the shapley importance should be computed.}

\item{target}{[\code{character(1)}] \cr
Only needed if \code{object} is not of class \code{WrappedModel} or \code{ResampleResult}.
Name of the target feature to be predicted.}

\item{bound.size}{[\code{numeric(1)}] \cr
Bound on the permutation size to compute the Shapley value (see Cohen et al. (2007)).}

\item{n.feat.perm}{[\code{numeric(1)}] \cr
The number of permutations of the feature(s) used to compute the feature importance.
The default is 50.}

\item{n.shapley.perm}{[\code{numeric(1)} | \code{"all.unique"}] \cr
The number of permutations that should be used for the shapley value.
Use \code{"all.unique"} to use all unique permutations.
If n.shapley.perm > number of all unique permutatios then only all unique permutations are used.
Default is 120.}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}} | \code{function} | list of \code{function}] \cr
Performance measure(s) used to measure the model performance.
Can also be a named list of function with signature \code{function(y, pred)},
where \code{y} and \code{pred} are vectors containing the true and the predicted values of the target.}

\item{predict.fun}{[\code{function}] \cr
Only needed if \code{object} is not of class \code{WrappedModel} or \code{ResampleResult}.
The signature must be \code{function(object, newdata)} and the function should always return a vector of predictions.
The default \code{NULL} internally uses \code{predict(object, newdata = newdata)}.}

\item{value.function}{[\code{function}] \cr
Function that defines the value function which is used to compute the shapley value.}
}
\description{
Computes the shapley importance of a feature.
}
\references{
Cohen, S., Dror, G., & Ruppin, E. (2007).
Feature selection via coalitional game theory.
Neural Computation, 19(7), 1939-1961.
}
